+++ src/GitHub.VisualStudio/UI/Views/Controls/LoginControl.xaml.cs
@@ -9,14 +9,18 @@ using GitHub.UI;
 using GitHub.ViewModels;
 using NullGuard;
 using ReactiveUI;
+using System.ComponentModel.Composition;
 
 namespace GitHub.VisualStudio.UI.Views.Controls
 {
+    public class GenericLoginControl : SimpleViewUserControl<ILoginControlViewModel, LoginControl>
+    { }
+
     /// <summary>
     /// Interaction logic for LoginControl.xaml
     /// </summary>
     [ExportView(ViewType=UIViewType.Login)]
-    public partial class LoginControl : SimpleViewUserControl, IViewFor<ILoginControlViewModel>, IView
+    public partial class LoginControl : GenericLoginControl
     {
         public LoginControl()
         {
@@ -109,28 +113,6 @@ namespace GitHub.VisualStudio.UI.Views.Controls
                 .Select(x => x == LoginMode.EnterpriseOnly)
                 .Where(x => x == true)
                 .BindTo(this, v => v.enterpriseTab.IsSelected));
-        }
-
-        public static readonly DependencyProperty ViewModelProperty = DependencyProperty.Register(
-            "ViewModel", typeof(ILoginControlViewModel), typeof(LoginControl), new PropertyMetadata(null));
-
-        object IViewFor.ViewModel
-        {
-            get { return ViewModel; }
-            set { ViewModel = (ILoginControlViewModel)value; }
-        }
-
-        object IView.ViewModel
-        {
-            get { return ViewModel; }
-            set { ViewModel = (ILoginControlViewModel)value; }
-        }
-
-        public ILoginControlViewModel ViewModel
-        {
-            [return: AllowNull]
-            get { return (ILoginControlViewModel)GetValue(ViewModelProperty); }
-            set { SetValue(ViewModelProperty, value); }
         }
     }
 }
