+++ src/GitHub.VisualStudio/UI/Views/Controls/RepositoryCreationControl.xaml
@@ -1,4 +1,4 @@
-﻿<uirx:SimpleViewUserControl x:Class="GitHub.VisualStudio.UI.Views.Controls.RepositoryCreationControl"
+﻿<local:GenericRepositoryCreationControl x:Class="GitHub.VisualStudio.UI.Views.Controls.RepositoryCreationControl"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
@@ -231,4 +231,4 @@
         </StackPanel>
      
     </DockPanel>
-</uirx:SimpleViewUserControl>
+</local:GenericRepositoryCreationControl>
