+++ src/GitHub.VisualStudio/UI/Views/Controls/RepositoryPublishControl.xaml
@@ -1,4 +1,4 @@
-﻿<uirx:SimpleViewUserControl x:Class="GitHub.VisualStudio.UI.Views.Controls.RepositoryPublishControl"
+﻿<local:GenericRepositoryPublishControl x:Class="GitHub.VisualStudio.UI.Views.Controls.RepositoryPublishControl"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
@@ -154,4 +154,4 @@
         </Button>
 
     </StackPanel>
-</uirx:SimpleViewUserControl>
+</local:GenericRepositoryPublishControl>
